Add all filters to side panel
